requirements.txt
flask
msal
requests
python-dotenv